+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Manage Secrets in the same namespace in which CertConfigs are found.
+
 ### Changed
 
 - Adjust helm chart to be used with `config-controller`.